refactored project directoru structure. Much cleaner now :)
