--- radsat-sk/common/common.h
@@ -1,13 +1,0 @@
-/*
- * common.h
- *
- *  Created on: Sep 28, 2020
- *      Author: Tyrel
- */
-
-#ifndef COMMON_H_
-#define COMMON_H_
-
-
-
-#endif /* COMMON_H_ */
