+++ radsat-sk/framework/fileio/RFram.h
@@ -1,0 +1,12 @@
+/**
+ * @file RFram.h
+ * @date
+ * @author
+ */
+
+#ifndef RFRAM_H_
+#define RFRAM_H_
+
+
+
+#endif /* RFRAM_H_ */
