+++ radsat-sk/framework/utils/RCommon.h
@@ -1,0 +1,13 @@
+/**
+ * @file RCommon.h
+ * @date
+ * @author
+ */
+
+#ifndef RCOMMON_H_
+#define RCOMMON_H_
+
+#include <stdint.h>
+
+
+#endif /* RCOMMON_H_ */
