+++ radsat-sk/framework/wrappers/RUart.h
@@ -1,0 +1,12 @@
+/**
+ * @file RUart.h
+ * @date
+ * @author
+ */
+
+#ifndef RUART_H_
+#define RUART_H_
+
+
+
+#endif /* RUART_H_ */
