+++ radsat-sk/operation/collection/RLogging.h
@@ -1,0 +1,12 @@
+/**
+ * @file RLogging.h
+ * @date
+ * @author
+ */
+
+#ifndef RLOGGING_H_
+#define RLOGGING_H_
+
+
+
+#endif /* RLOGGING_H_ */
