+++ radsat-sk/src/tasks/RPayloadCollectionTask.c
@@ -1,0 +1,7 @@
+/**
+ * @file RPayloadCollectionTask.c
+ * @date March 27, 2021
+ * @author Tyrel Kostyk
+ */
+
+
