+++ radsat-sk/src/tasks/RPayloadCollectionTask.h
@@ -1,0 +1,12 @@
+/**
+ * @file RPayloadCollectionTask.h
+ * @date March 27, 2021
+ * @author Tyrel Kostyk
+ */
+
+#ifndef RPAYLOADCOLLECTIONTASK_H_
+#define RPAYLOADCOLLECTIONTASK_H_
+
+
+
+#endif /* RPAYLOADCOLLECTIONTASK_H_ */
